requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
dnspython==1.16.0
ecdsa==0.18.0
fastapi==0.109.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
pytest==7.4.4
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
Random-Word==1.0.11
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
